requirements.txt
Flask==2.3.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
